+++ index.integration.test.js
@@ -1,5 +1,25 @@
 
 const request = require('supertest');
+
+const setupIntegrationTest = () => {
+    const { exec } = require("child_process");
+
+    exec("sqlite3 database-test.db < src/database/schema.sql", (error, stdout, stderr) => {
+        if (error) {
+            console.log(`error: ${error.message}`);
+            return;
+        }
+        if (stderr) {
+            console.log(`stderr: ${stderr}`);
+            return;
+        }
+        console.log(`stdout: ${stdout}`);
+    });
+};
+
+setupIntegrationTest();
+
+// setupIntegrationTest to be run first to ensure database is created before app
 const app = require("./index")
 
 /*