+++ jest.config.js
@@ -1,0 +1,2 @@
+process.env.DATABASE_NAME = 'database-test.db';
+